Added angular velocity / acceleration and changed all floating point numbers to floats

+++ src/definitions.h
@@ -1,10 +1,10 @@
 #pragma once
 #include "include.h"
 //PI to 4
-constexpr double PI = 3.1416;
+constexpr float PI = 3.1416;
 
 // m/s^2
-constexpr double earth_g = 9.81;
+constexpr float earth_g = 9.81;
 
 //orange color
 sf::Color orange1(242, 101, 34);
